[graphql] Replace `Int` with `BigInt` for overflow

+++ python_modules/dagster-graphql/dagster_graphql/schema/asset_checks.py
@@ -33,7 +33,7 @@ GrapheneAssetCheckExecutionResolvedStatus = graphene.Enum.from_enum(
 
 
 class GrapheneAssetCheckEvaluationTargetMaterializationData(graphene.ObjectType):
-    storageId = graphene.NonNull(graphene.Int)
+    storageId = graphene.NonNull(graphene.BigInt)
     runId = graphene.NonNull(graphene.String)
     timestamp = graphene.NonNull(graphene.Float)
